Lots of stuff
Renamed MainActivity to ThemeActivity to support the MainActivity in the
manifest. Updated the FragmentTheme to add all of the switch activity
like the contact to make cards launcher and action. Added theme actions
to manifest, adw class, icons class, icon provider to class. The theme
apply is non functional. Manifest and other associated items need to be
cleaned up