Major Overhaul and All Launchers Working
Moved all of the Fragmentx files into the fragment folder. Moved all of
the IconFragment files in the iconfragment folder. Moved the strings for
the different tabs in the ThemeActivity from strings.xml into their own
string_tabname.xml to keep things in better order. Go Launcher is now
fully functional as well as Smart Launcher. Currently Action, Go and
Smart will redirect to play if not installed, will add for the rest.
Also the Manifest now includes permission for write_external so
IconRequest is 100%. Changed FragmentText to FragmentAbout. Will expand
later.